requirements.txt
django
pillow
djangorestframework
django-filter
psycopg2
djmail
django-jinja
python-decouple
django-oauth-toolkit
djangorestframework-simplejwt

#pip3 install -r requirements.txt (run command to install these libraries)